Move around query helpers into OData module;
